requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
diffusers==0.23.1
docker-pycreds==0.4.0
einops==0.7.0
filelock==3.9.0
fire==0.5.0
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
huggingface-hub==0.19.4
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
lightning-utilities==0.10.0
lmdb==1.4.1
lpips==0.1.4
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.0
numpy==1.24.1
packaging==23.2
Pillow==9.3.0
piq==0.8.0
protobuf==4.25.1
psutil==5.9.6
pytorch-lightning==2.1.2
PyYAML==6.0.1
regex==2023.10.3
requests==2.28.1
safetensors==0.4.0
scipy==1.11.4
sentry-sdk==1.35.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
termcolor==2.3.0
-f https://download.pytorch.org/whl/torch_stable.html 
torch==2.1.1+cu118
torch-fidelity==0.3.0
torchaudio==2.1.1+cu118
torchmetrics==1.2.0
torchvision==0.16.1+cu118
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.4.0
urllib3==1.26.13
wandb==0.16.0
yarl==1.9.3
zipp==3.17.0
